+++ cnxarchive/tests/test_views.py
@@ -561,90 +561,6 @@ class ViewsTestCase(unittest.TestCase):
         resp_body = get_content_html(self.request).body
         self.assertTrue(resp_body.startswith('<html'))
 
-    def test_exports(self):
-        # Test for the retrieval of exports (e.g. pdf files).
-        id = 'e79ffde3-7fb4-4af3-9ec8-df648b391597'
-        version = '7.1'
-        type = 'pdf'
-        ident_hash = '{}@{}'.format(id, version)
-        filename = "{}@{}.{}".format(id, version, type)
-
-        # Build the request.
-        self.request.matchdict = {'ident_hash': ident_hash,
-                                  'type': type,
-                                  }
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'export'
-
-        from ..views import get_export
-        export = get_export(self.request).body
-
-        self.assertEqual(self.request.response.content_disposition,
-                         "attached; filename=college-physics-{}.pdf"
-                         .format(version))
-        expected_file = os.path.join(testing.DATA_DIRECTORY, 'exports',
-                                     filename)
-        with open(expected_file, 'r') as file:
-            self.assertEqual(export, file.read())
-
-        # Test exports can access the other exports directory
-        id = '56f1c5c1-4014-450d-a477-2121e276beca'
-        version = '8'
-        ident_hash = '{}@{}'.format(id, version)
-        filename = '{}@{}.pdf'.format(id, version)
-        self.request.matchdict = {'ident_hash': ident_hash,
-                                  'type': 'pdf'
-                                  }
-
-        export = get_export(self.request).body
-        self.assertEqual(
-            self.request.response.content_disposition,
-            "attached; filename=elasticity-stress-and-strain-{}.pdf"
-            .format(version))
-
-        expected_file = os.path.join(testing.DATA_DIRECTORY, 'exports2',
-                                     filename)
-        with open(expected_file, 'r') as file:
-            self.assertEqual(export, file.read())
-
-    def test_exports_type_not_supported(self):
-        # Build the request
-        self.request.matchdict = {
-                'ident_hash': '56f1c5c1-4014-450d-a477-2121e276beca@8',
-                'type': 'txt'
-                }
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'export'
-
-        from ..views import get_export
-        self.assertRaises(httpexceptions.HTTPNotFound,
-                          get_export, self.request)
-
-    def test_exports_404(self):
-        # Build the request
-        self.request.matchdict = {
-                'ident_hash': '24184288-14b9-11e3-86ac-207c8f4fa432@0',
-                'type': 'pdf'
-                }
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'export'
-
-        from ..views import get_export
-        self.assertRaises(httpexceptions.HTTPNotFound,
-                          get_export, self.request)
-
-    def test_exports_without_version(self):
-        id = 'ae3e18de-638d-4738-b804-dc69cd4db3a3'
-
-        # Build the request
-        self.request.matchdict = {'ident_hash': id, 'type': 'pdf'}
-        self.request.matched_route = mock.Mock()
-        self.request.matched_route.name = 'export'
-
-        from ..views import get_export
-        with self.assertRaises(IdentHashMissingVersion) as cm:
-            get_export(self.request)
-
     def test_get_extra_404(self):
         id = '94919e72-7573-4ed4-828e-673c1fe0cf9b'
         version = '66.1'
